+++ frontend/src/renderer/src/components/StatusPage.tsx
@@ -13,7 +13,7 @@ enum ConnectionStatus {
 function StatusPage(): JSX.Element {
   const [waco, set_waco] = useState<number>(()=>{
     const savedWaco = localStorage.getItem('balance')
-    return savedWaco ? parseInt(savedWaco) : 0
+    return savedWaco ? parseFloat(savedWaco) : 0
   })
   const [peers, set_peers] = useState<number>(0)
   const [status, set_status] = useState<ConnectionStatus>(ConnectionStatus.Connected);
